refactor: remove output port directory for category
